!!! TASK: Move workspace module to separate package
