+++ Neos.Neos/composer.json
@@ -12,7 +12,6 @@
     },
     "require": {
         "php": "^8.2",
-        "neos/diff": "self.version",
         "neos/flow": "~9.0.0",
         "neos/media-browser": "self.version",
         "neos/party": "~7.0.3",
@@ -24,8 +23,9 @@
         "neos/fusion-afx": "self.version",
         "neos/fusion-form": "^1.0 || ^2.0",
         "neos/fluid-adaptor": "~9.0.0",
-        "behat/transliterator": "~1.0",
-        "neos/media": "self.version"
+        "neos/media": "self.version",
+        "neos/workspace-ui": "self.version",
+        "behat/transliterator": "~1.0"
     },
     "suggest": {
         "neos/site-kickstarter": "Helps with creating new site packages for Neos.",
