+++ Neos.Workspace.Ui/Classes/Controller/WorkspacesController.php
@@ -12,7 +12,7 @@
 
 declare(strict_types=1);
 
-namespace Neos\Neos\Controller\Module\Management;
+namespace Neos\Workspace\Ui\Controller;
 
 use Doctrine\DBAL\DBALException;
 use Neos\ContentRepository\Core\ContentRepository;
