+++ Neos.Workspace.Ui/Classes/Controller/WorkspacesControllerInternals.php
@@ -12,7 +12,7 @@
 
 declare(strict_types=1);
 
-namespace Neos\Neos\Controller\Module\Management;
+namespace Neos\Workspace\Ui\Controller;
 
 use Neos\ContentRepository\Core\Dimension\ContentDimension;
 use Neos\ContentRepository\Core\Dimension\ContentDimensionSourceInterface;
