+++ Neos.Workspace.Ui/Configuration/Policy.yaml
@@ -1,0 +1,59 @@
+privilegeTargets:
+
+  'Neos\Flow\Security\Authorization\Privilege\Method\MethodPrivilege':
+
+    'Neos.Neos:Backend.PublishAllToLiveWorkspace':
+      label: Allowed to publish to the live workspace
+      matcher: 'method(Neos\Workspace\Ui\Controller\WorkspacesController->publishWorkspaceAction(workspace.baseWorkspace.name === "live"))'
+
+    'Neos.Neos:Backend.CreateWorkspaces':
+      label: Allowed to create a workspace
+      matcher: 'method(Neos\Workspace\Ui\Controller\WorkspacesController->(create|new)Action())'
+
+    'Neos.Neos:Backend.Module.Management.Workspaces.ManageOwnWorkspaces':
+      label: Allowed to manage own workspaces
+      matcher: 'method(Neos\Workspace\Ui\Controller\WorkspacesController->(publishWorkspace|discardWorkspace|edit|update|delete)Action(workspace.owner === current.userInformation.backendUser))'
+
+    'Neos.Neos:Backend.Module.Management.Workspaces.ManageInternalWorkspaces':
+      label: Manage internal workspaces
+      matcher: 'method(Neos\Workspace\Ui\Controller\WorkspacesController->(publishWorkspace|discardWorkspace|edit|update|delete)Action(workspace.owner === null))'
+
+    'Neos.Neos:Backend.Module.Management.Workspaces.ManageAllPrivateWorkspaces':
+      label: Manage all private workspaces
+      matcher: 'method(Neos\Workspace\Ui\Controller\WorkspacesController->(publishWorkspace|discardWorkspace|edit|update|delete)Action()) && evaluate(this.workspace.owner !== current.userInformation.backendUser, this.workspace.personalWorkspace === false)'
+
+  'Neos\Neos\Security\Authorization\Privilege\ModulePrivilege':
+    'Neos.Neos:Backend.Module.Management.Workspaces':
+      label: General access to the workspace module
+      matcher: 'management/workspaces'
+
+roles:
+  'Neos.Neos:LivePublisher':
+    privileges:
+      -
+        privilegeTarget: 'Neos.Neos:Backend.PublishAllToLiveWorkspace'
+        permission: GRANT
+
+  'Neos.Neos:AbstractEditor':
+    privileges:
+      -
+        privilegeTarget: 'Neos.Neos:Backend.CreateWorkspaces'
+        permission: GRANT
+
+      -
+        privilegeTarget: 'Neos.Neos:Backend.Module.Management.Workspaces.ManageOwnWorkspaces'
+        permission: GRANT
+
+      -
+        privilegeTarget: 'Neos.Neos:Backend.Module.Management.Workspaces'
+        permission: GRANT
+
+  'Neos.Neos:Administrator':
+    privileges:
+      -
+        privilegeTarget: 'Neos.Neos:Backend.Module.Management.Workspaces.ManageInternalWorkspaces'
+        permission: GRANT
+
+      -
+        privilegeTarget: 'Neos.Neos:Backend.Module.Management.Workspaces.ManageAllPrivateWorkspaces'
+        permission: GRANT
