+++ Neos.Workspace.Ui/Configuration/Settings.yaml
@@ -1,0 +1,11 @@
+Neos:
+  Neos:
+    modules:
+      management:
+        submodules:
+          workspaces:
+            label: 'Neos.Neos:Modules:workspaces.label'
+            controller: 'Neos\Workspace\Ui\Controller\WorkspacesController'
+            description: 'Neos.Neos:Modules:workspaces.description'
+            icon: fas fa-th-large
+            mainStylesheet: 'Lite'
