+++ Neos.Workspace.Ui/Resources/Private/Partials/Workspaces/ContentChangeAttributes.html
