+++ Neos.Workspace.Ui/Resources/Private/Partials/Workspaces/ContentChangeDiff.html
