+++ Neos.Workspace.Ui/Resources/Private/Templates/Workspaces/Edit.html
