+++ Neos.Workspace.Ui/Resources/Private/Templates/Workspaces/Index.html
