+++ Neos.Workspace.Ui/Resources/Private/Templates/Workspaces/New.html
