+++ Neos.Workspace.Ui/Resources/Private/Templates/Workspaces/Show.html
@@ -64,8 +64,8 @@
 										<td class="check neos-priority1">
 											<label for="{change.serializedNodeAddress}" class="neos-checkbox"><f:form.checkbox name="nodes[]" value="{change.serializedNodeAddress}" id="{change.serializedNodeAddress}" /><span></span></label>
 										</td>
-										<td id="change-{change.serializedNodeAddress}" {f:render(partial: 'Module/Management/Workspaces/ContentChangeAttributes', arguments: {change: change, nodepath: relativePath})} data-neos-toggle="tooltip" data-placement="left" data-container="body">
-											<f:render partial="Module/Management/Workspaces/ContentChangeDiff" arguments="{change: change, contentDimensions: contentDimensions}"/>
+										<td id="change-{change.serializedNodeAddress}" {f:render(partial: 'Workspaces/ContentChangeAttributes', arguments: {change: change, nodepath: relativePath})} data-neos-toggle="tooltip" data-placement="left" data-container="body">
+											<f:render partial="Workspaces/ContentChangeDiff" arguments="{change: change, contentDimensions: contentDimensions}"/>
 										</td>
 										<td class="neos-action">
 											<f:if condition="{canPublishToBaseWorkspace}">
