+++ Neos.Workspace.Ui/composer.json
@@ -1,0 +1,18 @@
+{
+    "name": "neos/workspace-ui",
+    "type": "neos-package",
+    "license": [
+        "GPL-3.0-or-later"
+    ],
+    "description": "",
+    "require": {
+        "php": "^8.2",
+        "neos/diff": "self.version"
+    },
+    "autoload": {
+        "psr-4": {
+            "Neos\\Workspace\\Ui\\": "Classes"
+        }
+    },
+    "extra": {}
+}
